mktables: Use correct structure to look up data
There are two types of tables in mktables: Map tables map code points
to the values a property have for those code points; and match tables
which are booleans, give "does a code point match a given property
value?".  There are different data structures to encapsulate each.  This
code was using the wrong structure to look something up.  Usually this
failed, and a fall-back value was used instead.  When compiling an early
Unicode release, I discovered that there could be a conflict.