regcomp.c, perl.c: Use 'VOL' not 'volatile'
I presume that VOL is so perl can at least compile on compilers that lack
'volatile'.  (It was added by Larry in 1991 by 352d5a3ab; there were
probably a lot more compilers like that back then.)  But VOL is used
consistently in the core, with these being the two exceptions.