Suggest having the previous release ready.
Each time I do a release, I find myself scrambling to download the
tarball of the previous release, and install it so it can be compared
to the release that's worked on. Added a paragraph as a reminder to
that before the day of the release.

(cherry picked from commit c282dc5ebef25e8b002cd913928f847fab339716)