Revert "include SvREADONLY() in SvIMMORTAL() test"
This reverts commit 3041a168eaf8027c9e376a53450bcb825f527da8.

.text section of VC Win32 perl523.dll threaded -O1 before is 0xE3DDB bytes
of machine code long, after is 0xE3DBB. Instructions are saved because an
extra branch was removed. A check of a ptr in a register against a constant
(PL_* global on unthreaded) or near constant (reg with my_perl+offset)
avoids a memory read of the SV head to get the flags, for the previous
reasons and to enable a future optimization to SvIMMORTAL, revert this
commit.