remove SvREADONLY && SvIMMORTAL cargo culting
Prior to commit d689ffdd6d in 5.004, there was some sketchy inconsistent
behavior of using RO flag to indicate immortals. The RO flag on the SV
is really irrelavent to immortals, since we know there are only 4 immortal
SVs possible. Adding more at runtime is impossible. An immortal without
RO flag is impossible and illegal/mem corruption if it happens. Checking
the RO flag involves a mem read of the SV head and branch. To simplify
the code, remove the RO flag check branch and just do the bounds checks
in SvIMMORTAL, no memory read involved since var sv will already be a
register probably. This commit is releated to commit "simplify SvIMMORTAL"
and "Revert "include SvREADONLY() in SvIMMORTAL() test" and commit
3041a168ea .

.text section of VC Win32 perl523.dll threaded -O1 before is 0xE3D1B bytes
of machine code long and after this commit is 0xE3CFB.