Document e57270be442bfaa9dc23eebd67485e5a806b44e3:
I wasn't sure where or how much of it to document, but it seems
like it's important, and this relating to permissions not being
removed, I consider it security-related. This is similiar to what
the original Debian ticket that relates to it mentioned.

I've cut Niko's text a bit shorter, taken from the commit message
itself.

(cherry picked from commit 07450df75585c03dda7c25f07a88139918aeb0ea)