embed.fnc: Alter 'b' flag meaning
This commit changes this flag to mean that the backward compatibility
functions are compiled unless the -DNO_MATHOMS cflag is specified to
Configure.  Previously the meaning was sort of like that but not
precisely.

Doing this means that the prototypes that needed to be manually added to
mathoms.c are no longer needed.  No special parameter assertions have to
be made.  makedef.pl no longer needs to parse mathoms.c and have special
cases for it.  And several special case entries in embed.fnc can be
non-special cased.