bisect-runner: Only run ./Configure -S when needed
We only need to run it when config.sh has been modified, which only
happens before 5.10.  From v5.23.4-46-g41d7307 onwards ./Configure -S
did not work, until v5.23.5-90-g473edb6, so we need to skip this if we
want to bisect in that region.  (Or we could just skip the die, but we
might as well skip the system call, too.)